Issue #1530 - fhir-term refactoring
